migrations

+++ database/migrations/2019_06_03_000008_create_realty_objects_table.php
@@ -12,11 +12,11 @@ class CreateRealtyObjectsTable extends Migration
             $table->increments('id');
             $table->date('planned_contact')->nullable();
             $table->string('cadastral_numb')->nullable();
-            $table->float('area', 4, 1);
+            $table->decimal('area', 4, 1);
             $table->integer('power')->nullable();
-            $table->float('ceiling', 4, 2)->nullable();
+            $table->decimal('ceiling', 4, 2)->nullable();
             $table->string('contract_status');
-            $table->float('commission', 4, 1)->nullable();
+            $table->decimal('commission', 4, 1)->nullable();
             $table->longText('description')->nullable();
             $table->decimal('cost', 15, 2)->nullable();
             $table->decimal('cost_m', 15, 2)->nullable();
